+++ Ketchup/Model/Post.swift
@@ -28,11 +28,11 @@ struct Post: Identifiable, Codable {
     var fromInAppCamera: Bool
     var repost: Bool
     var didRepost: Bool
-    var overallRating: Int
-    var serviceRating: Int
-    var atmosphereRating: Int
-    var valueRating: Int
-    var foodRating: Int
+    var overallRating: Double
+    var serviceRating: Double
+    var atmosphereRating: Double
+    var valueRating: Double
+    var foodRating: Double
     var coordinates: CLLocationCoordinate2D? {
         if let point = self.restaurant.geoPoint {
             return CLLocationCoordinate2D(latitude: point.latitude, longitude: point.longitude)}
@@ -58,11 +58,11 @@ struct Post: Identifiable, Codable {
         self.fromInAppCamera = try container.decode(Bool.self, forKey: .fromInAppCamera)
         self.repost = try container.decodeIfPresent(Bool.self, forKey: .repost) ?? false
         self.didRepost = try container.decodeIfPresent(Bool.self, forKey: .didRepost) ?? false
-        self.overallRating = try container.decode(Int.self, forKey: .overallRating)
-        self.serviceRating = try container.decode(Int.self, forKey: .serviceRating)
-        self.atmosphereRating = try container.decode(Int.self, forKey: .atmosphereRating)
-        self.valueRating = try container.decode(Int.self, forKey: .valueRating)
-        self.foodRating = try container.decode(Int.self, forKey: .foodRating)
+        self.overallRating = try container.decode(Double.self, forKey: .overallRating)
+        self.serviceRating = try container.decode(Double.self, forKey: .serviceRating)
+        self.atmosphereRating = try container.decode(Double.self, forKey: .atmosphereRating)
+        self.valueRating = try container.decode(Double.self, forKey: .valueRating)
+        self.foodRating = try container.decode(Double.self, forKey: .foodRating)
         
     }
 
@@ -83,11 +83,11 @@ struct Post: Identifiable, Codable {
         fromInAppCamera: Bool,
         repost: Bool = false,
         didRepost: Bool = false,
-        overallRating: Int,
-        serviceRating: Int,
-        atmosphereRating: Int,
-        valueRating: Int,
-        foodRating: Int
+        overallRating: Double,
+        serviceRating: Double,
+        atmosphereRating: Double,
+        valueRating: Double,
+        foodRating: Double
     ) {
         self.id = id
         self.mediaType = mediaType
